(chore) move LogDetail component hooks into /hooks
