+++ client/src/components/logbooks/LogDetail/ItemSection.tsx
@@ -1,7 +1,7 @@
+import useItemSection from "@/components/logbooks/LogDetail/hooks/useItemSection";
 import ItemRowsTable from "@/components/logbooks/LogDetail/ItemRowsTable";
 import { getRowsForItem } from "@/components/logbooks/LogDetail/lib/get-rows";
 import { Button } from "@/components/logbooks/LogDetail/style/_common.style";
-import useItemSection from "@/components/logbooks/LogDetail/useItemSection";
 import NewItem from "@/components/logbooks/NewItem/NewItem";
 import Modal from "@/components/utility/Modal/Modal";
 import type { Item, ItemTemplate } from "@t/data/logbook.types";
