+++ client/src/components/logbooks/LogDetail/LogDetail.tsx
@@ -1,6 +1,6 @@
+import useLogDetail from "@/components/logbooks/LogDetail/hooks/useLogDetail";
 import ItemSection from "@/components/logbooks/LogDetail/ItemSection";
 import { getSectionsForLog } from "@/components/logbooks/LogDetail/lib/get-sections";
-import useLogDetail from "@/components/logbooks/LogDetail/useLogDetail";
 import type { ID } from "@t/data/utility.types";
 import S from "./style/LogDetail.style";
