+++ client/src/components/logbooks/LogDetail/hooks/useNewItemRow.ts
